@@ -1,10 +1,12 @@
 # frozen_string_literal: true
 
+require 'active_support/core_ext/integer/time'
+
 Rails.application.configure do
   # Settings specified here will take precedence over those in config/application.rb.
 
-  # In the development environment your application's code is reloaded on
-  # every request. This slows down response time but is perfect for development
+  # In the development environment your application's code is reloaded any time
+  # it changes. This slows down response time but is perfect for development
   # since you don't have to restart the web server when you make code changes.
   config.cache_classes = false
 
@@ -44,6 +46,12 @@ Rails.application.configure do
   # Print deprecation notices to the Rails logger.
   config.active_support.deprecation = :log
 
+  # Raise exceptions for disallowed deprecations.
+  config.active_support.disallowed_deprecation = :raise
+
+  # Tell Active Support which deprecation messages to disallow.
+  config.active_support.disallowed_deprecation_warnings = []
+
   # Raise an error on page load if there are pending migrations.
   config.active_record.migration_error = :page_load
 
@@ -51,7 +59,10 @@ Rails.application.configure do
   config.active_record.verbose_query_logs = true
 
   # Raises error for missing translations.
-  # config.action_view.raise_on_missing_translations = true
+  # config.i18n.raise_on_missing_translations = true
+
+  # Annotate rendered view with file names.
+  # config.action_view.annotate_rendered_view_with_filenames = true
 
   # Use an evented file watcher to asynchronously detect changes in source code,
   # routes, locales, etc. This feature depends on the listen gem.
@@ -72,7 +83,7 @@ Rails.application.configure do
     )
   end
 
-  # Perform jobs straight away 
+  # Perform jobs straight away
   config.active_job.queue_adapter = :inline
 
   config.after_initialize do
@@ -83,10 +94,12 @@ Rails.application.configure do
     Bullet.add_footer = true
     Bullet.skip_html_injection = false
 
-    Bullet.add_whitelist :type => :unused_eager_loading, class_name: "Lesson", association: :topic
-    Bullet.add_whitelist :type => :unused_eager_loading, class_name: "Classroom", association: :subject
-    Bullet.add_whitelist :type => :unused_eager_loading, class_name: "ActionText::RichText", :association => :embeds_attachments
-    Bullet.add_whitelist :type => :unused_eager_loading, class_name: "Customisation", :association => :image_attachment
+    Bullet.add_whitelist type: :unused_eager_loading, class_name: 'Lesson', association: :topic
+    Bullet.add_whitelist type: :unused_eager_loading, class_name: 'Classroom', association: :subject
+    Bullet.add_whitelist type: :unused_eager_loading, class_name: 'ActionText::RichText', association: :embeds_attachments
+    Bullet.add_whitelist type: :unused_eager_loading, class_name: 'Customisation', association: :image_attachment
   end
 
+  # Uncomment if you wish to allow Action Cable access from any origin.
+  # config.action_cable.disable_request_forgery_protection = true
 end
