@@ -14,8 +14,8 @@ Rails.application.routes.draw do
       patch 'sync'
     end
   end
-  resources :leaderboard, only:[:show, :index]
-  resources :classrooms, only: [:show, :index, :update]
+  resources :leaderboard, only: %i[show index]
+  resources :classrooms, only: %i[show index update]
   resources :questions do
     collection do
       get 'topic'
@@ -33,27 +33,27 @@ Rails.application.routes.draw do
   resources :topics
   resources :subjects
   resources :homeworks
-  resources :users, only:[:show, :index, :update] do
-      member do
-        patch 'set_role'
-        patch 'reset_password'
-        patch 'update_email'
-        post 'send_welcome_email'
-        delete 'remove_role'
-        delete 'unlink_oauth_account'
-      end
-      collection do
-        get 'manage_roles'
-      end
+  resources :users, only: %i[show index update] do
+    member do
+      patch 'set_role'
+      patch 'reset_password'
+      patch 'update_email'
+      post 'send_welcome_email'
+      delete 'remove_role'
+      delete 'unlink_oauth_account'
+    end
+    collection do
+      get 'manage_roles'
+    end
   end
-  resources :admins, only:[:show] do
+  resources :admins, only: [:show] do
     member do
       post 'become'
       post 'reset_year'
     end
   end
 
-  resources :flagged_questions, only:[:create]
+  resources :flagged_questions, only: [:create]
   resources :school_groups
   resources :lessons
   resources :customisations do
@@ -68,7 +68,7 @@ Rails.application.routes.draw do
   get 'quizzes/new/:subject', to: 'quizzes#new'
   get 'dashboard/', to: 'dashboard#show'
 
-  get "/pages/*id", to: 'pages#show', as: :page, format: false
+  get '/pages/*id', to: 'pages#show', as: :page, format: false
 
   authenticated :user do
     root to: 'dashboard#show', as: :authenticated_root
